+++ packages/react-components-lab/src/components/Card/Card.stories.tsx
@@ -6,6 +6,20 @@ import Card from './Card';
 import { CardProps } from './types';
 import productImage from './product.png';
 
+export default {
+  title: 'Example/Card',
+  component: Card,
+  argTypes: {
+    title: { type: 'string' },
+    subTitle: { type: 'array' },
+    description: { type: 'string' },
+    isOpen: { type: 'boolean' },
+    image: { type: 'string' },
+    isClickable: { type: 'boolean' },
+    disabled: { type: 'boolean' },
+  },
+} as Meta;
+
 const Template: Story<CardProps> = (args) => {
   const [isActive, setIsActive] = useState(false);
 
@@ -59,7 +73,19 @@ WithImage.args = {
   ),
 };
 
-export default {
-  title: 'Example/Card',
-  component: Card,
-} as Meta;
+export const Disabled = Template.bind({});
+Disabled.args = {
+  title: 'Disabled',
+  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit',
+  image: productImage as string,
+  disabled: true,
+  isOpen: true,
+  children: (
+    <Box>
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+      tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+      veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
+      commodo consequat.
+    </Box>
+  ),
+};